+++ src/currency.spec.ts
@@ -6,6 +6,8 @@ const parts = formatter.formatToParts(1000);
 const { value: g } = parts[1];
 
 describe("format", () => {
+	const badNum = Number.POSITIVE_INFINITY;
+
 	describe("default", () => {
 		test("positive", () => {
 			expect(Currency.format(1)).toBe("0.01");
@@ -16,6 +18,13 @@ describe("format", () => {
 		test("zero", () => {
 			expect(Currency.format(0)).toBe("0.00");
 			expect(Currency.format(-0)).toBe("0.00");
+		});
+		test("undefined", () => {
+			expect(Currency.format(badNum)).toBeUndefined();
+		});
+		test("extreme", () => {
+			expect(Currency.format(Number.MAX_SAFE_INTEGER))
+				.toBe(`90${g}071${g}992${g}547${g}409.91`);
 		});
 	});
 
@@ -30,12 +39,27 @@ describe("format", () => {
 			expect(Currency.format(0, true)).toBe("0.00");
 			expect(Currency.format(-0, true)).toBe("0.00");
 		});
+		test("undefined", () => {
+			expect(Currency.format(badNum, true)).toBeUndefined();
+		});
 	});
 
-	test("extreme", () => {
-		expect(Currency.format(Number.MAX_SAFE_INTEGER))
-			.toBe(`90${g}071${g}992${g}547${g}409.91`);
+	describe("invert", () => {
+		test("positive", () => {
+			expect(Currency.format(1, false, true)).toBe("-0.01");
+		});
+		test("negative", () => {
+			expect(Currency.format(-1, false, true)).toBe("0.01");
+		});
+		test("zero", () => {
+			expect(Currency.format(0, false, true)).toBe("0.00");
+			expect(Currency.format(-0, false, true)).toBe("0.00");
+		});
+		test("undefined", () => {
+			expect(Currency.format(badNum, false, true)).toBeUndefined();
+		});
 	});
+
 });
 
 describe("String -> Number", () => {
@@ -65,6 +89,7 @@ describe("String -> Number", () => {
 	});
 
 	test("full - large number", () => {
+		expect(Currency.strToNum("1234567.89")).toBe(123456789);
 		expect(Currency.strToNum("1,234,567.89")).toBe(123456789);
 	});
 
@@ -110,8 +135,7 @@ describe("Number -> String", () => {
 	});
 
 	test("large number", () => {
-
-		expect(Currency.numtoStr(123456789)).toBe(`1${g}234${g}567.89`);
+		expect(Currency.numtoStr(123456789)).toBe(`1234567.89`);
 	});
 
 });